fixed pointers are not permitted as case values
./sys/sys/kern_sig.c:472:3: error: pointers are not permitted as case values
  472 |   case SIG_DFL:
      |   ^~~~
./sys/sys/kern_sig.c:515:3: error: pointers are not permitted as case values
  515 |   case SIG_IGN:

+++ sys/sys/kern_sig.c
@@ -469,7 +469,7 @@ issignal(p)
 
 		switch ((int)u.u_signal[sig]) {
 
-		case SIG_DFL:
+		case (int)SIG_DFL:
 			/*
 			 * Don't take default actions on system processes.
 			 */
@@ -512,7 +512,7 @@ issignal(p)
 				return(sig);
 			/*NOTREACHED*/
 
-		case SIG_IGN:
+		case (int)SIG_IGN:
 			/*
 			 * Masking above should prevent us
 			 * ever trying to take action on a held
